Fix benchmark suite: isolate WAMR build, split wasmtime/wasmer TUs, fix buffer reuse
- WAMR built in isolated subdirectory to prevent global compile flag pollution
- wasmtime and wasmer in separate .cpp files to avoid wasm-c-api header conflicts
- WAMR/wasmtime/wasmer are mutually exclusive (all export wasm-c-api symbols)
- wasm3 can be combined with any single one of the above
- Fix WAMR target detection (AARCH64 not arm64)
- Fix wasmer to use static library
- Copy WASM bytes for wasm3/WAMR (both modify input buffers)

Verified: all 4 runtimes build and run correctly on ARM64.
